+++ mobile/src/core/CAText.js
@@ -1,6 +1,13 @@
 import React from 'react';
 import { StyleSheet, Text } from 'react-native';
 
+/**
+ * Returns a resusable Text component used by screens
+ * @param {boolean} props.appColor if text is to be displayed in app's color theme
+ * @param {object} props.style style object to be applied to button
+ * @param {string} props.size specified size of the text
+ * @param {string} props.children actual text to be rendered
+ */
 export default function CAText({ appColor, style, size, children }) {
   const applyAppColor = appColor ? styles.appColor : null;
   return (