+++ mobile/src/screens/Attendee/Contact.js
@@ -3,6 +3,13 @@ import { StyleSheet, View } from 'react-native';
 
 import CAText from '@core/CAText';
 
+/**
+ * Returns the Contact screen
+ * 
+ * @param {object} props.route React Navigation route object containing information passed from previous screen
+ * 
+ * This screen presents selected sponsor contact information
+ */
 export default function Contact({ route }) {
   const { 
     name,
@@ -26,7 +33,8 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     justifyContent: 'center',
-    alignItems: 'center'
+    alignItems: 'center',
+    backgroundColor: '#fff'
   },
   text: {
     paddingBottom: 20