+++ mobile/src/screens/Attendee/PresentQR.js
@@ -3,23 +3,24 @@ import { Image, StyleSheet, View } from 'react-native';
 
 import CAText from '@core/CAText';
 import Spacer from '@core/Spacer';
+import appText from '@utils/text';
 import testQR from '@images/testQR.png';
 
-const text = {
-  qrHeader: 'Your QR Code',
-  qrMessage: 'Expect to be asked to present your QR Code when checking in, getting food/snacks, and winning prizes!'
-};
-
+/**
+ * Returns the PresentQR screen
+ * 
+ * This screen presents an attendee's QR Code for the purpose of verifying identification
+ */
 export default function PresentQR() {
   return (
     <View style={styles.container}>
-      <CAText appColor size='xlg'>{text.qrHeader}</CAText>
+      <CAText appColor size='xlg'>{appText.qrHeader}</CAText>
       <Spacer size='xlg'/>
       <Image source={testQR} style={styles.displayQR}/>
       <Spacer size='xlg'/>
       <View style={styles.msgContainer}>
-        <CAText style={{ color: '#A9A9A9' }} size='xsm'>
-          {text.qrMessage}
+        <CAText style={styles.subtitle} size='xsm'>
+          {appText.qrMessage}
         </CAText>
       </View>
     </View>
@@ -38,5 +39,8 @@ const styles = StyleSheet.create({
   },
   msgContainer: {
     width: '80%'
+  },
+  subtitle: {
+    color: '#A9A9A9'
   }
 });