+++ mobile/src/screens/User/Profile.js
@@ -5,15 +5,17 @@ import { FontAwesome5 } from '@expo/vector-icons';
 import CAText from '@core/CAText';
 import CAButton from '@core/CAButton';
 
+import appText from '@utils/text';
+
 import data from '@data/data.json';
 
-const text = {
-  conferenceID: 'Conference ID: ',
-  register: 'You are registered as a(n): ',
-  loadNewConferenceTitle: 'Load New Conference',
-  loadNewConferenceMsg: 'You are using the app with the above conference loaded. You can change this by loading a new conference that you are registered to!'
-}
-
+/**
+ * Returns the Profile screen
+ * 
+ * @param {object} props.navigation React Navigation navigation obtaining allowing for traversal to different screens
+ * 
+ * Displays the currently logged in user's contact information as well as the currently loaded conference instance
+ */
 export default function Profile({ navigation }) {
   const [profileData, setProfileData] = useState({});
   const [instanceData, setInstanceData] = useState({});
@@ -38,20 +40,20 @@ export default function Profile({ navigation }) {
       <View style={styles.horizontalRule} />
       <View>
         <CAText size='sm' style={styles.description}>
-          {text.conferenceID}
+          {appText.conferenceID}
           <CAText size="sm" style={styles.bolded}>
             {instanceData.id}
           </CAText>
         </CAText>
         <CAText size='sm' style={styles.description}>
-          {text.register}
+          {appText.register}
           <CAText size='sm' style={styles.bolded}>
             {instanceData.role}
           </CAText>
         </CAText>
-        <CAButton style={styles.loadButton} title={text.loadNewConferenceTitle} size='sm'/>
-        <CAText style={{ alignSelf: 'center', color: '#A9A9A9'}} size='xsm'>
-            {text.loadNewConferenceMsg}
+        <CAButton style={styles.loadButton} title={appText.loadNewConferenceTitle} size='sm'/>
+        <CAText style={styles.loadMessage} size='xsm'>
+            {appText.loadNewConferenceMsg}
           </CAText>
       </View>
       <View style={styles.bottomActionContainer}>
@@ -97,6 +99,11 @@ const styles = StyleSheet.create({
   loadButton: {
     margin: 30,
     backgroundColor: '#A9A9A9'
+  },
+  loadMessage: {
+    alignSelf: 'center',
+    color: '#A9A9A9',
+    width: '80%'
   },
   logoutButton: {
     marginBottom: 20,