+++ scripts/faker/faker.js
@@ -10,6 +10,8 @@ var data = {
   users: [],
 }
 
+// Custom data point generating functions
+
 const getTicketType = () => {
   const num = faker.random.number(2);
   return ['Standard', 'Full-Access', 'VIP'][num];
@@ -17,9 +19,10 @@ const getTicketType = () => {
 
 const getApplicationStatus = () => {
   const num = faker.random.number(2);
-  return ['Accepted', 'Pending', 'Confirmed'][num];
+  return ['Accepted', 'Pending', 'Rejected'][num];
 }
 
+// Genereate data and append to data.users above
 
 for (let i = 0; i < 2000; i++) {
   // Feel free to ammend this USER type as needed during development