+++ website/src/App.js
@@ -3,6 +3,10 @@ import { Banner } from "@components/Banner";
 import { Dashboard } from "@admin/Dashboard";
 import { Applicants } from "@admin/Applicants";
 import { BrowserHome } from "@browser/BrowserHome";
+import { Products } from "@product/Products"
+import { About } from "@product/About";
+import { Registrations } from "@admin/Applicants";
+import { Statistics } from "@admin/Statistics";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 
 import './App.css';
@@ -13,7 +17,10 @@ function App() {
       <BrowserRouter>
         <Banner/>
         <Switch>
-          <Route path="/admin/applications"><Applicants/></Route>
+          <Route path="/products"><Products/></Route>
+          <Route path="/about"><About/></Route>
+          <Route path="/admin/registrations"><Registrations/></Route>
+          <Route path="/admin/statistics"><Statistics/></Route>
           <Route path="/admin"><Dashboard/></Route>
           <Route path="/browser"><BrowserHome/></Route>
           <Route path="/"><Home /></Route>