+++ website/src/components/admin/Dashboard/index.js
@@ -1,36 +1,46 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Container, Grid, Typography } from "@material-ui/core";
 import { Sidebar } from "@admin/common/Sidebar";
 import { NumberWidget } from "@admin/common/NumberWidget";
 import { NavigationPanel } from "./NavigationPanel";
+import { getAttendees } from "@utils/data";
 import { PieChartWidget } from "@admin/common/PieChartWidget";
 import "./dashboard.css";
 
 /**
  * Component for the /admin landing page
- * See README for figma links to mockup design
- * @param {*} props
+ * See README for figma links to mockup design and purpose
  */
 export const Dashboard = (props) => {
 
   const applicationsData = {
-    labels: ['Pending', 'Accepted'],
+    labels: ['Pending', 'Accepted', 'Rejected'],
     datasets: [
       {
-        label: '# of Votes',
-        data: [156, 680],
+        data: [680, 156, 78],
         backgroundColor: [
           'rgba(255, 206, 86, 0.2)',
-          'rgba(75, 192, 192, 0.2)'
+          'rgba(75, 192, 192, 0.2)',
+          'rgba(255, 99, 132, 0.2)'
         ],
         borderColor: [
           'rgba(255, 206, 86, 1)',
-          'rgba(75, 192, 192, 1)'
-        ],
-        borderWidth: 1,
-      },
-    ],
-  }
+          'rgba(75, 192, 192, 1)',
+          'rgba(255, 99, 132, 1)'],
+          borderWidth: 1,
+        },
+      ],
+    }
+  const [attendees, setAttendees] = useState([]);
+  
+  const getData = async () => {
+    const { users: rows } = await getAttendees();
+    setAttendees(rows);
+  };
+
+  useEffect(() => {
+    getData();
+  }, []);
 
   return (
     <Container className="dashboard" maxWidth="lg">