+++ website/src/components/admin/common/NumberWidget.jsx
@@ -8,9 +8,11 @@ import {
 import './common.css';
 
 /**
- * Generic component to display live feed of an integer based statistics
+ * Generic component to display integer based statistics
  * (i.e. the current total attendees registered for an event)
- * @param {*} props 
+ * @param {string} props.value The number to display in the integer data widget
+ * @param {string} props.heading The label for this data point
+ * @param {object} props.data The statistical data and formatting for the chart.
  */
 export const NumberWidget = (props) => {
 
@@ -18,8 +20,8 @@ export const NumberWidget = (props) => {
     <div className="number-widget">
       <Card className="number-widget-card">
         <CardContent>
-          <Typography color="textSecondary" component="h6" variant="h6">{props.heading}</Typography>
-          <Typography color="textPrimary" component="h2" variant="h2">
+          <Typography color="textSecondary" component="h7" variant="h7">{props.heading}</Typography>
+          <Typography color="textPrimary" component="h3" variant="h3">
             {props.value}
           </Typography>
         </CardContent>