+++ CHANGELOG.md
@@ -13,16 +13,38 @@ e.g
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
-
+## Mobile
 ### Added
 - Added mobile screen to sign in ([57e3adf](https://github.com/coffeexcode/capstone/commit/57e3adff75627cc3015e0de6aea4cfafbdf00df0))
 - Added mobile screen to Present QR ([75f72c2](https://github.com/coffeexcode/capstone/commit/75f72c232a4990a185aac5d14c8294165ca6d739))
 - Added mobile screen to Scan QR codes ([0f56699](https://github.com/coffeexcode/capstone/commit/0f56699db40f170264c931d6813e8c88da64d26d))
 - Added mobile screen to view the schedule of a conference ([4ffa223](https://github.com/coffeexcode/capstone/commit/4ffa2238b79ee4b5540eabbee2a511331ff47d6f))
 - Added mobile screen to view profile ([#1](https://github.com/coffeexcode/capstone/pull/1))
-- Added website page for managing event applications or registrations ([#4](https://github.com/coffeexcode/capstone/pull/4))
 - Added organizer/sponsor/attendee views with their respective functions via bottom tab bar ([#3](https://github.com/coffeexcode/capstone/pull/3))
 - Added About page to view conference information and sponsors ([#5](https://github.com/coffeexcode/capstone/pull/5))
+- Added statistics page to display graphs and visual statistics ([#9](https://github.com/coffeexcode/capstone/pull/9))
+- Added event page to allow users to register/withdraw from event offerings ([#10](https://github.com/coffeexcode/capstone/pull/10))
+
+### Changed
+
+## Website
+### Added
+- Added landing page for administrative section of the website. ([#2](https://github.com/coffeexcode/capstone/pull/2))
+- Added website page for managing event applications or registrations ([#4](https://github.com/coffeexcode/capstone/pull/4))
+- Added website pages for ConAssist team details and offering page ([#13](https://github.com/coffeexcode/capstone/pull/13))
+- Added page to view events ([bfbcc43](https://github.com/coffeexcode/capstone/pull/14))
+- Added page to view conferences ([bfbcc43](https://github.com/coffeexcode/capstone/pull/14))
+- Added button to register for conference ([bfbcc43](https://github.com/coffeexcode/capstone/pull/14))
+- Added button to set a reminder for conference ([bfbcc43](https://github.com/coffeexcode/capstone/pull/14))
+- Added button to sponsor a conference ([bfbcc43](https://github.com/coffeexcode/capstone/pull/14))
+
+### Changed
+
+## Backend
+### Added
+- Added REST endpoints to create and interact with user accounts ([699a6c5](https://github.com/Jailoodu/RestfulRegistration/commit/699a6c5e41b2a884352639cd666e45d9bb4bc58a))
+- Added REST endpoint to review conference applications ([699a6c5](https://github.com/Jailoodu/RestfulRegistration/commit/699a6c5e41b2a884352639cd666e45d9bb4bc58a))
+- Dockerized application so it can be run in various environments   ([757f2e5](https://github.com/Jailoodu/RestfulRegistration/commit/757f2e5eec1960ba19bdb70b576f9b9d6a6c20bc))
+- Added REST endpoint to export event attendee data  ([dd20591](https://github.com/Jailoodu/RestfulRegistration/commit/dd20591ee28f3af258593a111bdee0e1ac33640c))
 
 ### Changed